+++ packages/web/controls/__stories__/TextInputInteractive.stories.tsx
@@ -1,7 +1,10 @@
+/* eslint-disable @typescript-eslint/await-thenable */
 import React from 'react';
 import { expect } from '@storybook/jest';
 import { ComponentStoryObj } from '@storybook/react';
 import { userEvent, waitFor, within } from '@storybook/testing-library';
+
+import { pauseStory } from '../../utils/storybook';
 
 import { AddressForm } from './AddressForm';
 
@@ -18,12 +21,24 @@ export const FormFilled: ComponentStoryObj<typeof AddressForm> = {
   },
   play: async ({ canvasElement }) => {
     const canvas = within(canvasElement);
-    userEvent.type(canvas.getByPlaceholderText('4321 Jade Palace'), '221 Stockholm Street');
-    userEvent.type(canvas.getByLabelText('City/town'), 'San Jose');
-    userEvent.type(canvas.getByLabelText('State'), 'California');
-    userEvent.type(canvas.getByLabelText('Postal code'), '89201');
-    userEvent.type(canvas.getByLabelText('Country'), 'United States');
-    userEvent.click(canvas.getByTestId('save-btn'));
+    await userEvent.type(canvas.getByPlaceholderText('4321 Jade Palace'), '221 Stockholm Street');
+    await pauseStory(1000);
+
+    await userEvent.type(canvas.getByLabelText('City/town'), 'San Jose');
+    await pauseStory(1000);
+
+    await userEvent.type(canvas.getByLabelText('State'), 'California');
+    await pauseStory(1000);
+
+    await userEvent.type(canvas.getByLabelText('Postal code'), '89201');
+    await pauseStory(1000);
+
+    await userEvent.type(canvas.getByLabelText('Country'), 'United States');
+    await pauseStory(1000);
+
+    await userEvent.click(canvas.getByTestId('save-btn'));
+    await pauseStory(1000);
+
     await waitFor(() => expect(canvas.getByText('Submit button was clicked')).toBeInTheDocument());
   },
 };